Refactor package structure: move `OpenInBDVCommand` to `plugins` package and update imports accordingly
