+++ src/test/java/sc/fiji/ome/zarr/pyramid/Pyramidal5DImageDataDemo.java
@@ -2,6 +2,8 @@ package sc.fiji.ome.zarr.pyramid;
 
 import net.imagej.ImageJ;
 
+import sc.fiji.ome.zarr.plugins.OpenInBDVCommand;
+
 public class Pyramidal5DImageDataDemo
 {
 	public static void main( String[] args )
